sending REST API request in objective C in iOS
There are  lots of requirements where calling the specific API on the web is a must. For example, email API, SMS API, database API, server restart API, stock market API, etc. 

While  the sendSynchronous method is deprecated, iOS supports the NSURLsession to send the request to the server asynchronously. This code will also help to post the query data into the server by wrapping it into a request object.