Making Asynchronous request synchronous in iOS
There are lot of requirements while building iOS app where you need to fetch the data from the server beforehand or perform some operation after certain data is fetched from the server.

Since sendSynchronous request is deprecated and the new NSURLSession call is Asynchronous, we need to perform some sort of wait operation till the Asynchronous request is completed. - Semaphore is the solution.